Added semantic annotation support to view builder
