+++ src/labthings/server/semantics/moz.py
@@ -1,18 +1,7 @@
-from .. import decorators
 from .. import fields
+from .base import Property
 
 # BASIC PROPERTIES
-class Property:
-    def __init__(self, schema):
-        self.schema = schema
-
-    def __call__(self, viewcls):
-        # Use the class name as the semantic type
-        viewcls = decorators.Semtype(self.__class__.__name__)(viewcls)
-        viewcls = decorators.PropertySchema(self.schema)(viewcls)
-        return viewcls
-
-
 class BooleanProperty(Property):
     """https://iot.mozilla.org/schemas/#BooleanProperty"""
